general folder restructure
