+++ src/components/authentication/LoginForm.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { Link } from "react-router-dom";
-import { connect } from "react-redux";
+import { connect } from 'react-redux';
 import { withFormik, Form, Field } from "formik";
 import * as Yup from "yup";
-import { register } from "../actions/index";
+import { login } from '../../actions/index';
 import logo from "../globals/design-elements/logo.png";
 import arrow from "../images/left-arrow.png";
 import "../index.css";
@@ -15,21 +15,20 @@ import {
   H1,
   DescriptionForm,
   LinkFooter
-} from "../globals/form-styles";
+} from "../../globals/form-styles";
 
 function FormTemplate({ touched, errors }) {
   return (
     <MainDiv>
       <ArrowDiv>
         <Link to='/'>
-        {/* <Link to="/"> */}
           <img src={arrow} alt="arrow" />
         </Link>
       </ArrowDiv>
       <Logo>
         <img src={logo} alt="Logo" />
       </Logo>
-      <H1> Register</H1>
+      <H1> Login</H1>
       <DescriptionForm>
         {" "}
         Create a profile, follow your favorites cooks, save delicious recipes
@@ -37,10 +36,6 @@ function FormTemplate({ touched, errors }) {
       </DescriptionForm>
 
       <Form className="formik-form">
-        <div>
-          {touched.email && errors.email}
-          <Field type="email" name="email" class="input" placeholder="Email" />
-        </div>
         <div>
           {touched.username && errors.username}
           <Field
@@ -50,7 +45,7 @@ function FormTemplate({ touched, errors }) {
             placeholder="Username"
           />
         </div>
-        <div>
+        <div className="spaced-div">
           {touched.password && errors.password}
           <Field
             type="password"
@@ -59,39 +54,31 @@ function FormTemplate({ touched, errors }) {
             placeholder="Password"
           />
         </div>
-
         <button type="submit" className="register-button">
-          Register
+          Login
         </button>
-
-        <FooterDiv>
-          <p>
-            Already have an account?
-            <Link to="/login" style={{ textDecoration: "none" }}>
-              {""}
-              <LinkFooter>Log in</LinkFooter>
-            </Link>
-          </p>
-        </FooterDiv>
       </Form>
+      <FooterDiv>
+        <p>
+          Don't have an account?{" "}
+          <Link to="/register" style={{ textDecoration: "none" }}>
+            <LinkFooter> Register</LinkFooter>
+          </Link>{" "}
+        </p>
+      </FooterDiv>
     </MainDiv>
   );
 }
-const FormikRegisterForm = withFormik({
-  mapPropsToValues({ email, username, password }) {
+const FormikLoginForm = withFormik({
+  mapPropsToValues({ username, password }) {
     return {
-      email: email || "",
       username: username || "",
       password: password || ""
     };
   },
-  // Validation //
   validationSchema: Yup.object().shape({
-    email: Yup.string()
-      .email("An email is required")
-      .required("Please, add an email"),
     username: Yup.string().required(
-      " Username is required to have acess to the kitchen"
+      "Username is required to have acess to the kitchen"
     ),
     password: Yup.string()
       .min(
@@ -100,9 +87,12 @@ const FormikRegisterForm = withFormik({
       )
       .required("A password is required to have access to the kitchen")
   }),
-  handleSubmit(values, { props }) {
-    props.register(values, props.history);
+
+  handleSubmit(values, {props} ) {
+    props.login(values, props.history)
   }
 })(FormTemplate);
 
-export default connect(state => state, { register })(FormikRegisterForm);
+export default connect(
+  state => state, { login }
+) (FormikLoginForm)
