+++ src/components/createRecipe/RecipeFormFour.js
@@ -2,7 +2,7 @@ import React from "react";
 import styles from "styled-components";
 import back from "../globals/design-elements/back.png";
 import check from "../globals/design-elements/check.png";
-import Footer from "./Footer";
+import Footer from "../navigation/FooterNav";
 import CheckIcon from '@material-ui/icons/Check';
 import ArrowBackIcon from '@material-ui/icons/ArrowBack';
 import { TextField, Select, MenuItem } from '@material-ui/core';
@@ -17,7 +17,7 @@ import {
   Title2,
   AddItem,
   Step
-} from "../globals/form-styles";
+} from "../../globals/form-styles";
 
 export default function RecipeFormFour (props) { 
   const { history, step, prevPage, title, nextPage, onHandleChange, onHandleSubmit, onHandleInstructionInput, onAddInstruction, instruction, instructions, onHandleFinalSubmit } = props;
