+++ src/components/instructions/Instruction.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { connect } from 'react-redux'
-import * as actionCreators from '../actions/index'
+import * as actionCreators from '../../actions/index'
 
 const Instruction = ({ instruction }) => { 
   return (
