+++ src/components/recipe/RecipeView.js
@@ -1,10 +1,10 @@
 import React, { useEffect } from "react";
 import { connect } from "react-redux";
-import * as actionCreators from "../actions/index";
+import * as actionCreators from "../../actions/index";
 import Recipe from "./Recipe";
 import SearchBar from "./SearchBar";
 import Footer from "./Footer";
-import Loader from "./Loader";
+import Loader from "../Loader";
 import "../App.css";
 
 const RecipeView = ({ getRecipes, recipeView, isFetching }) => {
